+++ Pandora/Paradigm/Inventory/Ability.hs
@@ -1,1 +1,5 @@
-module Pandora.Paradigm.Inventory.Ability where
+module Pandora.Paradigm.Inventory.Ability (module Exports) where
+
+import Pandora.Paradigm.Inventory.Ability.Modifiable as Exports
+import Pandora.Paradigm.Inventory.Ability.Replaceable as Exports
+import Pandora.Paradigm.Inventory.Ability.Viewable as Exports